Add files via upload
This is the output of Assignment 1 submitted by me on Sunday 01/27/2019. I forgot to post the screenshot of the output the same day so I am doing this on 01/29/2019. I assure you that no changes have been made to the program. I request you to please consider this submission of the screenshot.